sk/python/sk.ui copy/requirements.txt
streamlit
azure-storage-blob
requests
python-dotenv
azure-identity